telegram-bot/pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["Jorik Schellekens <joriksch@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.0"
python-dotenv = "^0.19.0"
litellm = "^1.52.3"
flask = "^3.0.3"
dill = "^0.3.9"
eth-account = "^0.13.4"
eth-typing = "^5.0.1"
web3 = "^7.5.0"
psutil = "^6.1.0"
sentry-sdk = "^2.18.0"
prometheus-client = "^0.21.0"
googlesearch-python = "^1.2.5"
beautifulsoup4 = "^4.12.3"
types-beautifulsoup4 = "^4.12.0.20241020"
types-psutil = "^6.1.0.20241102"
swarmer = {path = "../swarmer"}
pre-commit = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.18.0"
pytest-cov = "^3.0"
mypy = "^1.8.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
markers = [
    "asyncio: mark test as async",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
